requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.1
ddgs>=9.0.0
wikipedia>=1.4.0
pytest>=8.3.0
